requirements.txt
alembic==1.16.2
blinker==1.9.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
XlsxWriter==3.2.3
python-magic==0.4.27
flask-migrate
